Add component tests

+++ src/domain/audience.py
@@ -2,6 +2,6 @@ from pydantic import BaseModel
 
 class Audience(BaseModel):
 	size: int
-	executive: float
-	professionals: float
-	other: float
+	executive: int
+	professionals: int
+	other: int
